% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_volcano_plot.R
\name{ctk.volcano}
\alias{ctk.volcano}
\title{Generate Volcano Plot}
\usage{
ctk.volcano(
  df,
  var_x,
  var_y,
  Symbol = NULL,
  title = NULL,
  subtitle = NULL,
  x_axis_label = NULL,
  x_limits = NULL,
  y_axis_label = NULL,
  y_limits = NULL,
  FC_cut = c(-1, 1),
  FC_cut_line = T,
  Q_cut = 0.05,
  Q_cut_line = T,
  pt_color = c("firebrick3", "navy", "grey", "orange"),
  pt_size = 2,
  pt_alpha = 0.5,
  Highlight = NULL,
  legend = F,
  label = T,
  label_size = 3,
  overlaps = 50,
  ticks = c(5, 5),
  minor_ticks = c(1, 1),
  na_omit = T
)
}
\arguments{
\item{df}{A data frame containing the differential expression data.}

\item{var_x}{A character string specifying the column name for the x-axis variable (e.g., log2 fold change).}

\item{var_y}{A character string specifying the column name for the y-axis variable (e.g., p-value adjust).}

\item{Symbol}{An optional character string specifying the column name for gene symbols or identifiers.}

\item{title}{An optional character string specifying the main title of the plot.}

\item{subtitle}{An optional character string specifying the subtitle of the plot.}

\item{x_axis_label}{An optional character string specifying the label for the x-axis.}

\item{x_limits}{A numeric vector of length 2, specifying the limits for the x-axis.}

\item{y_axis_label}{An optional character string specifying the label for the y-axis.}

\item{y_limits}{A numeric vector of length 2, specifying the limits for the y-axis.}

\item{FC_cut}{A numeric vector of length 2, specifying the cut-off values for fold change.}

\item{FC_cut_line}{A logical value indicating whether to draw dashed lines for fold change cut-offs.}

\item{Q_cut}{A numeric value specifying the cut-off value for the q-value or adjusted p-value.}

\item{Q_cut_line}{A logical value indicating whether to draw a dashed line for the q-value cut-off.}

\item{pt_color}{A character vector of length 4, specifying the colors for up-regulated, down-regulated, non-significant, and highlighted points.}

\item{pt_size}{A numeric value specifying the size of the points.}

\item{pt_alpha}{A numeric value between 0 and 1, specifying the transparency of the points.}

\item{Highlight}{An optional vector specifying the gene symbols or identifiers to be highlighted, or a numeric value specifying the number of top up/down-regulated genes to be highlighted.}

\item{legend}{A logical value indicating whether to show the legend.}

\item{label}{A logical value indicating whether to label the highlighted points.}

\item{label_size}{A numeric value specifying the size of the labels for highlighted points.}

\item{overlaps}{A numeric value specifying the maximum number of overlapping labels allowed.}

\item{ticks}{A numeric vector of length 2, specifying the number of ticks for the x and y axes.}

\item{minor_ticks}{A numeric vector of length 2, specifying the number of minor ticks for the x and y axes.}

\item{na_omit}{A logical value indicating whether to remove rows with missing values.}
}
\value{
A ggplot object representing the volcano plot.
}
\description{
This function generates a volcano plot for visualizing differential expression
data, with various customization options.
}
\examples{
# Generate a simple volcano plot
ctk.volcano(df, var_x = "log2FoldChange", var_y = "pvalue")

# Generate a volcano plot with customized settings
ctk.volcano(df, var_x = "log2FC", var_y = "padj",
            x_axis_label = "Log2 Fold Change",
            y_axis_label = "-Log10 Adjusted P-value",
            FC_cut = c(-1, 1), Q_cut = 0.05,
            Highlight = c("GENE1", "GENE2", "GENE3"),
            label = TRUE, label_size = 4)

}
