% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function_scatter_plot.R
\name{ctk.scatter}
\alias{ctk.scatter}
\title{This function is used to create a scatter plot and calculate correlation.}
\usage{
ctk.scatter(
  df,
  var_x,
  var_y,
  title = NULL,
  subtitle = NULL,
  x_axis_label = NULL,
  x_limits = NULL,
  y_axis_label = NULL,
  y_limits = NULL,
  legend_position = NULL,
  pt_size = 2,
  cor = "lm",
  cor_method = "pearson",
  cor_line_col = "firebrick2",
  cor_se = T,
  cor_se_col = "gray"
)
}
\arguments{
\item{df}{Data frame.}

\item{var_x}{Name of the variable for x-axis.}

\item{var_y}{Name of the variable for y-axis.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{x_axis_label}{Label for x-axis.}

\item{x_limits}{Limits for x-axis.}

\item{y_axis_label}{Label for y-axis.}

\item{y_limits}{Limits for y-axis.}

\item{legend_position}{Position of the legend.}

\item{pt_size}{Size of the points.}

\item{cor}{Whether to calculate correlation. Default is "lm", can be "FALSE" to avoid calculation.}

\item{cor_method}{Method for calculating correlation. Default is "pearson". One of "pearson", "kendall", or "spearman", can be abbreviated.}

\item{cor_line_col}{Color for correlation line.}

\item{cor_se}{Whether to show standard error for correlation line. Default is TRUE.}

\item{cor_se_col}{Color for standard error of correlation line.}
}
\value{
Returns a ggplot2 plot object.
}
\description{
This function is used to create a scatter plot and calculate correlation.
}
\examples{
library(Cplot)
df <- data.frame(x = rnorm(100), y = rnorm(100))
Cplot.scatter(df, "x", "y", title = "Scatter Plot", subtitle = "Sample Data")
}
